+++ app/src/main/java/ru/otus/basicarchitecture/model/FieldValueDto.kt
@@ -1,0 +1,6 @@
+package ru.otus.basicarchitecture.model
+
+interface FieldValueDto<T> {
+    val fValue: T
+    val isValid: Boolean
+}
