+++ app/src/main/java/ru/otus/basicarchitecture/model/LongFiledDto.kt
@@ -1,0 +1,3 @@
+package ru.otus.basicarchitecture.model
+
+class LongFiledDto(override val fValue: Long, override val isValid: Boolean) : FieldValueDto<Long?>
