+++ app/src/main/java/ru/otus/basicarchitecture/model/StringFiledDto.kt
@@ -1,0 +1,4 @@
+package ru.otus.basicarchitecture.model
+
+class StringFiledDto(override val fValue: String, override val isValid: Boolean) :
+    FieldValueDto<String>
