+++ app/src/main/java/ru/otus/basicarchitecture/model/ValidationEvent.kt
@@ -1,0 +1,10 @@
+package ru.otus.basicarchitecture.model
+
+sealed interface ValidationEvent {
+    object InvalidName : ValidationEvent
+    object ValidName : ValidationEvent
+    object InvalidSurname : ValidationEvent
+    object ValidSurname : ValidationEvent
+    object InvalidAge : ValidationEvent
+    object ValidAge : ValidationEvent
+}
