+++ app/src/main/java/ru/otus/basicarchitecture/view_model/TagsViewModel.kt
@@ -1,7 +1,10 @@
 package ru.otus.basicarchitecture.view_model
 
 import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import jakarta.inject.Inject
 
-class TagsViewModel : ViewModel() {
+@HiltViewModel
+class TagsViewModel @Inject constructor(): ViewModel() {
     // TODO: Implement the ViewModel
 }
